pyproject.toml
[project]
name = "moe-5g-nrx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
packages = [{ include = "src" }]
dependencies = [
    "matplotlib>=3.9.4",
    "numpy>=1.26.4",
    "hydra-core>=1.3.2",
    "sionna==0.19.2",
    "tensorflow-macos==2.15.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "tensorflow==2.15.0; sys_platform != 'darwin' or platform_machine != 'arm64'",
    "tf-keras==2.15.0",
    "torch>=2.8.0",
    "tqdm>=4.67.3",
    "wandb>=0.25.1",
]

[dependency-groups]
dev = ["ruff>=0.15.6", "pre-commit>=3.6.0"]

[tool.ruff]
line-length = 120
target-version = "py310"
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"
